refactor: reorganize balance tracker files and consolidate interfaces into a single location
